make setup dialog part of main window
