+++ css/passwordManager.css
@@ -1,18 +1,17 @@
 /* common - TODO add a shared stylesheet for this */
 
-body,
-html {
-	margin: 0;
-	padding: 0;
-	width: 100%;
-	height: 100%;
-	font-size: 16px;
-	font-family: '.SFNSText-Regular', 'Helvetica Neue', 'Arial', sans-serif;
-	overflow: auto;
+#bitwarden-setup-dialog {
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    background: #fff;
+    border-radius: 8px;
+    padding: 1em;
+		box-shadow: 0px 0px 5px rgba(0, 0, 0, 0.25);
+		z-index: 999;
 }
-* {
-	box-sizing: border-box;
-}
+
 .fade {
 	opacity: 0.333;
 }
