+++ index.html
@@ -17,6 +17,7 @@
 	<link rel="stylesheet" href="css/bookmarkManager.css"/>
 	<link rel="stylesheet" href="css/findinpage.css" />
 	<link rel="stylesheet" href="css/downloadManager.css" />
+	<link rel="stylesheet" href="css/passwordManager.css" />
 	<link rel="stylesheet" href="ext/font-awesome-4.4.0/css/font-awesome.min.css" />
 	<link rel="stylesheet" href="node_modules/dragula/dist/dragula.min.css" />
 </head>
@@ -106,6 +107,27 @@
 		<div id="download-container" role="list"></div>
 		<button id="download-close-button" class="fa fa-times"></button>
 	</div>
+
+	<div class="container padding" id="bitwarden-setup-dialog" hidden>
+		<h3 data-string="bitwardenSetupHeading"></h3>
+		<div class="light-fade" data-string="bitwardenSetupDescription" data-allowHTML></div>
+		<div class="spacer"></div>
+		<div class="light-fade" data-string="bitwardenSetupStep1" data-allowHTML></div>
+		<div class="spacer"></div>
+		<div class="light-fade" data-string="bitwardenSetupStep2" data-allowHTML></div>
+		<div class="spacer"></div>
+		
+		<div class="drag-and-drop-box" data-string="bitwardenDragBox">
+		</div>
+	
+		<div class="spacer"></div><div class="spacer"></div>
+	
+		<div class="buttons">
+		  <input type="button" id="bitwarden-setup-disable" data-value="bitwardenDisableAutofill"/>
+		  <input type="button" id="bitwarden-setup-cancel" data-value="bitwardenNotNow"/>
+		</div>
+	
+	  </div>
 
 	<!-- add scripts in Gruntfile.js -->
