+++ js/autofillSetup.js
@@ -1,4 +1,6 @@
-function checkAutofillSettings() {
+var showBitwardenDialog = require('passwordManager/bitwardenSetup.js')
+
+function checkAutofillSettings () {
   getActivePasswordManager().then((manager) => {
     if (!manager) {
       return { manager: null, configured: false }
@@ -10,18 +12,14 @@ function checkAutofillSettings() {
   }).then((result) => {
     const { manager, configured } = result
     if (manager && !configured) {
-      ipc.send('autofill-setup', { manager : manager.name }) 
+      if (manager.name === 'Bitwarden') {
+        showBitwardenDialog()
+      }
     }
   }).catch((err) => {
     console.error(err)
   })
 }
-
-// This handler will check autofill/password settings again after setup 
-// is finished. There's a possibility that it will fail, which will trigger
-// the setup dialog again. Thus it is important for a setup dialog to have
-// a 'disable' option, allowing user to end this cycle...
-ipc.on('password-autofill-reload', checkAutofillSettings)
 
 settings.listen('passwordManager', function (manager) {
   if (manager) {
