--- pages/autofillSetup/Bitwarden/index.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="UTF-8" />
-
-  <!-- title is set in JS -->
-  <link rel="stylesheet" href="../../../css/base.css" />
-  <link rel="stylesheet" href="bitwarden.css" />
-  <link rel="stylesheet" href="../../../ext/font-awesome-4.4.0/css/font-awesome.min.css" />
-</head>
-<body>
-  <div class="container padding">
-    <h3 data-string="bitwardenSetupHeading"></h3>
-    <div class="light-fade" data-string="bitwardenSetupDescription" data-allowHTML></div>
-    <div class="spacer"></div>
-    <div class="light-fade" data-string="bitwardenSetupStep1" data-allowHTML></div>
-    <div class="spacer"></div>
-    <div class="light-fade" data-string="bitwardenSetupStep2" data-allowHTML></div>
-    <div class="spacer"></div>
-    
-    <div class="drag-and-drop-box" data-string="bitwardenDragBox">
-    </div>
-
-    <div class="spacer"></div><div class="spacer"></div>
-
-    <div class="buttons">
-      <input type="button" id="disable" data-value="bitwardenDisableAutofill" onclick="disable()"/>
-      <input type="button" id="cancel" data-value="bitwardenNotNow" onclick="cancel()"/>
-    </div>
-
-  </div>
-
-  <script src="../../../ext/Dexie.min.js"></script>
-  <script src="../../../js/util/process.js"></script>
-  <script src="../../../dist/localization.build.js"></script>
-  <script src="bitwarden.js"></script>
-</body>
-
